requirements.txt
gen3_tracker>=0.0.4rc36
click
pandas
numpy
openpyxl
PyYAML
orjson
ndjson
inflection
jinja2
